firmware: add microcode package for Intel
Compiling the Intel microcode package results
in a microcode.bin and a microcode-64.bin. As we can
decide based on the subtarget which should be used,
we'll only split the required .bin file with
iucode-tool.

Instead of using the large microcode.bin/microcode-64.bin, the
splitted ucode files (separate for CPU families) will be
installed.

The microcodes are updated from preinit - that's the
earliest place where we can do it.

Use the latest 20180108 release, including:
 - update microcodes for various CPUs
 - Implements IBRS/IBPB support and enhances LFENCE: mitigation
   against Spectre (CVE-2017-5715)
 - downgrade microcodes for two platforms, due to reasons unknown
   from Intel

Signed-off-by: Zoltan HERPAI <wigyori@uid0.hu>