Fix broken and potentially paralyzing parallelism MAKEFLAGS
Really, any build system should use the makeflags passed to it.  The
exception is when such a system makes more than one sub-package like
OpenWRT.  In this case, we now have --load-average to over-loading CPU,
cache and memory, but this probably did not exist when OpenWRT's
parallelism was added.

But as-is it isn't just broken, it's horrific.  I stepped away and came
back to a five minute load average of 139 when I only have 4 CPU cores!
The OOM killer saved me.  This patch adds an alternative feature,
enabling the user to explicitly specify per-package makeflags.